requirements.txt
fastapi==0.100.0
httpx==0.24.1
openai==0.28.0
pydantic==2.1.1
redis[asyncio]==4.5.4
python-dotenv==1.0.0
uvicorn==0.23.2
